Fix crash with async main and no conc module
The concurrency module gets imported implicitly when
`-enable-experimental-concurrency` is passed to the compiler. If the
concurrency module isn't imported, we crash on th assert verifying that
we were able to find it. We need the module to lookup the
`_runAsyncMain` function. This patch updates the compiler to stop
type-checking the async-main function and emit an error if the module
isn't available instead of crashing.