RawSILInstLowering: delete an unused partial_apply when lowering assign_by_wrapper
So far we left this cleanup to SILCombine.
But the unused partial_apply (i.e. the "set" in case it's an init or the "init" in case it's a set) violates memory lifetime rules in case "self" is an inout.
Once we verify that, it would result in a memory lifetime violation.