MandatoryCopyPropagation support for unowned remnant lifetimes.
This supports debugging references between the time their canonical
lifetimes end and their scoped lifetime ends by creating an unowned
remnant during this time. The debugger will report that the object has
been destroyed, but it's type will be available, and the inline
contents can be inspected.

Adds an 'unownedRemnantMode' to CanonicalizeOSSALifetime. This
canonicalizes lifetimes exactly as before. That is important for
consistent behavior between debug and release builds. During
canonicalization, it does a little extra book-keeping to efficiently
track destroys that participate in the unowned remnant lifetime. After
canonicalization and unowned copy of the original value is created,
and these destroys are rewritten to destroy the unowned copy instead.

TODO: support unowned remnants for references wrapped in structs/tuples.

rdar://74544568 (-Onone compiler support for early object deinitialization)