Work Around +asserts Canonicalization Differences in Substitution
On +asserts builds, this code

public struct MemberInits<Value : Equatable> {
  fileprivate var value: String = "default"
}

extension MemberInits {
  public init<T>(_ array: [MemberInits<T>]) where Value == [T] {}
}

Produces an apply of the memberwise initializer substituting BoundGenericType,
on -asserts builds it produces a substitution of an ArraySliceType. This
suggests we have an assertion that is mutating the substitution map
somewhere. For now, switch to an explicit BoundGenericType to homogenize
this test while I'm looking into why this happens.