Updating tests for @completionHandlerAsync
This patch updates the tests for the @completionHandlerAsync attribute.

The main test is broken into three chunks; parsing, typechecking, and
the decl assginment phases.

For parsing, we verify both happy cases, with an explicitly provided
completion handler index and one where it is omitted. We also verify the
various ways that parsing can fail, with ill-formed names, missing
parens etc....

The typechecking phase verifies that the function the attribute is
attached to is not async, that it takes a completion handler, that the
completion handler is a function type, is a function that returns Void,
is @escaping, and is not @autoclosure.

The decl assignment complains about the name not referencing an existing
function, and the ambiguous naming of async functions.