[ABI] Renamed method descriptor's SignedAsData to IsAsync.
In 26b35494e655970a041c4a5f2085f16d82feeb05, new bits were spoken for in
MethodDescriptorFlags and ProtocolRequirementFlags to indicate that a
field is an async function pointer and as such must is treated as data
for the purposes of ptrauth.

Previously, that bit was named IsSignedAsData because indeed async
function pointers must be signed and authed as data.

We should make the representation semantic, however, so here that name
is changed to IsAsync.